refactor: Simplify input data.table for add_plots
The input data.table now has a 'x' column which contains the plot data
as either a filepath to the data, an object containing the data, or a
list with the server and uuid. Type is now derived from the new 'x'
column, source is derived from type.

These changes should make it much easier to add plots to the DB.