fix: Cohort pair name extraction
The pair name was being extracted from the basename of the pipeline
paths.

However, some of the pipeline paths (e.g signatures/.*/sig_inputs/.*)
do not have their basename as the sample name (it is a parent directory
higher up on the directory tree). Since the intention here was
to ensure correct correspondence between pair name and its data files
in the Cohort object (via the data.table merge), this has been replaced
with a more manual method, while still allowing for sample names to not
be the immediate basename.

TODO: decide whether to make the initial_dt non-optional instead of
initializing an empty default. The pair names should come from the
samplesheet, which should represent the ultimate source of truth (since
the pipeline output paths should all be generated using the pair names
provided in the samplesheet).